requirements.txt
fastapi>=0.115.0
uvicorn>=0.29.0
anthropic>=0.25.0
google-genai>=1.0.0
openai>=1.0.0
pypdf>=4.0.0
pdf2image>=1.17.0
pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
pytest>=8.0.0
httpx>=0.27.0
